Changed working directory, Changed location of Libraries to Load
Changed working directory as I changed the folder (as there was a typo in the original one and it could have got confusing) when we created a new clone and got rid of the fork. Also added the all libraries to be loaded at top instead of in between as you had advised.